+++ .idea/misc.xml
@@ -5,9 +5,9 @@
       <map>
         <entry key="app/src/main/res/drawable/progress_bar_drawable.xml" value="0.17" />
         <entry key="app/src/main/res/drawable/tortoise.xml" value="0.1395" />
-        <entry key="app/src/main/res/layout/activity_achievements.xml" value="0.2138364779874214" />
+        <entry key="app/src/main/res/layout/activity_achievements.xml" value="0.30616607666015627" />
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.28050231933593756" />
-        <entry key="app/src/main/res/layout/item_achievement.xml" value="0.27807971014492755" />
+        <entry key="app/src/main/res/layout/item_achievement.xml" value="0.1" />
         <entry key="app/src/main/res/mipmap-anydpi-v26/ic_launcher.xml" value="0.1395" />
         <entry key="app/src/main/res/mipmap-anydpi-v26/ic_launcher_round.xml" value="0.1395" />
         <entry key="app/src/main/res/xml/progress_bar.xml" value="0.2947916666666667" />
