+++ app/src/main/java/com/example/notlonesomegeorge/MainActivity.kt
@@ -16,8 +16,6 @@ class MainActivity : AppCompatActivity() {
     //DECLARE VARIABLES
     private lateinit var achBTN: ImageButton
     private lateinit var skillBTN: ImageButton
-    private lateinit var achRecyclerView: RecyclerView
-
 
     var mainActivityTag = "mainActivityTag";
 
@@ -29,11 +27,6 @@ class MainActivity : AppCompatActivity() {
         // INITIATE VARIABLES
         achBTN = findViewById(R.id.tortoiseBTN)
         skillBTN = findViewById(R.id.hatBTN)
-
-        // SET UP RECYCLER VIEW FOR ACHIEVEMENTS
-        achRecyclerView = findViewById(R.id.achievements_recycler)
-        achRecyclerView.layoutManager = LinearLayoutManager(this)
-        achRecyclerView.adapter = AchievementAdapter(this, getAchievementList())
 
         // DIRECT USER TO ACHIEVEMENTS ACTIVITY
         achBTN.setOnClickListener{
@@ -47,21 +40,6 @@ class MainActivity : AppCompatActivity() {
         Log.i(mainActivityTag, "In call_ach()...")
         val intent = Intent(this, AchievementsActivity::class.java)
         startActivity(intent)
-    }
-
-    // PROVIDE DATA SOURCE FOR ACHIEVEMENT LIST
-    private fun getAchievementList(): List<Achievement> {
-
-        var name: String
-        val achievements = ArrayList<Achievement>()
-
-        // create 100 homework!
-        for (i in 1..100){
-            name = getString(R.string.friend_tortoise_name)
-            achievements.add(Achievement(name))
-        }
-
-        return achievements
     }
 
 }
