+++ app/src/main/res/layout/item_achievement.xml
@@ -1,7 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
+    android:layout_height="wrap_content"
+    android:clipToPadding="false"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools">
 
@@ -20,6 +21,9 @@
         <androidx.constraintlayout.widget.ConstraintLayout
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
+            android:paddingBottom="8dp"
+            android:paddingLeft="8dp"
+            android:paddingRight="8dp"
             android:background="@color/olive_super_light_yellow">
 
             <LinearLayout
@@ -42,11 +46,11 @@
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:layout_weight="5"
-                    android:text="@string/friend_tortoise_name"
                     android:textAlignment="center"
                     android:textSize="17dp"
                     android:textColor="@color/olive_dark_green"
-                    android:layout_gravity="center"/>
+                    android:layout_gravity="center"
+                    tools:text="@string/friend_tortoise_name"/>
 
                 <Button
                     android:id="@+id/friend_description_btn"
@@ -54,7 +58,9 @@
                     android:layout_height="wrap_content"
                     android:layout_weight="1"
                     android:layout_gravity="center"
-                    android:text="@string/friend_description_BTN" />
+                    app:layout_constraintEnd_toEndOf="parent"
+                    app:layout_constraintHorizontal_bias="0"
+                    tools:text="@string/friend_description_BTN" />
 
             </LinearLayout>
